requirements.txt
transformers>=4.42.0
torch>=2.1.0
Pillow>=9.5.0
